Create taiyounionfind.py
Here is my union find algorithm for finding connected components, not great but it works. I don't know how to also upload txt file to go with it to store the graph. So it still needs that. I really REALLY don't want to just type the graph into my code (though I could because my example graph is small). To me this is important since storing a specific very large graph directly in your code would be intractable to just open and look at and edit your code. If anyone knows how to what I am saying in GitHub that would be very helpful!