Create graphAdjArrLine.txt
This is the graph adjacency array that goes with "taiyounionfind.py'. If you put it in the same folder as taiyounionfind.py and run it it wil work. I don't know how to make that make sense in GitHub, but I imagine it should be a normal thing since most actual programs use several different files that access others, and sometimes even other programming languages. When I look it up it sounds very hard, but I am sure someone can help me do that.